Move MethodImplementation trait to runtime module
