+++ crates/test-ui/ui/fn_ptr_reference_method.stderr
@@ -7,14 +7,14 @@ error[E0277]: the trait bound `for<'a> extern "C" fn(_, _, &'a NSObject): Method
   |                 required by a bound introduced by this call
   |
   = help: the following other types implement trait `MethodImplementation`:
-            unsafe extern "C" fn(Allocated<T>, objc2::runtime::Sel) -> __IdReturnValue
-            extern "C" fn(Allocated<T>, objc2::runtime::Sel) -> __IdReturnValue
-            unsafe extern "C" fn(Allocated<T>, objc2::runtime::Sel, A) -> __IdReturnValue
-            extern "C" fn(Allocated<T>, objc2::runtime::Sel, A) -> __IdReturnValue
-            unsafe extern "C" fn(Allocated<T>, objc2::runtime::Sel, A, B) -> __IdReturnValue
-            extern "C" fn(Allocated<T>, objc2::runtime::Sel, A, B) -> __IdReturnValue
-            unsafe extern "C" fn(Allocated<T>, objc2::runtime::Sel, A, B, C) -> __IdReturnValue
-            extern "C" fn(Allocated<T>, objc2::runtime::Sel, A, B, C) -> __IdReturnValue
+            unsafe extern "C" fn(Allocated<T>, objc2::runtime::Sel) -> IdReturnValue
+            extern "C" fn(Allocated<T>, objc2::runtime::Sel) -> IdReturnValue
+            unsafe extern "C" fn(Allocated<T>, objc2::runtime::Sel, A) -> IdReturnValue
+            extern "C" fn(Allocated<T>, objc2::runtime::Sel, A) -> IdReturnValue
+            unsafe extern "C" fn(Allocated<T>, objc2::runtime::Sel, A, B) -> IdReturnValue
+            extern "C" fn(Allocated<T>, objc2::runtime::Sel, A, B) -> IdReturnValue
+            unsafe extern "C" fn(Allocated<T>, objc2::runtime::Sel, A, B, C) -> IdReturnValue
+            extern "C" fn(Allocated<T>, objc2::runtime::Sel, A, B, C) -> IdReturnValue
           and $N others
 note: required by a bound in `ClassBuilder::add_method`
  --> $WORKSPACE/crates/objc2/src/declare/mod.rs
